feat(ui): redesign v2 — money strip on top, wide pace charts stay
Per Eric's review of v1: keep the big side-by-side 5h/7d charts
(PaceChartCard restored verbatim from main, minus its freshness chip
which now lives in the page header) and give the top strip a different
job instead.

The strip becomes the money story: Today + a wide This-month tile that
absorbs the bottom-of-page MonthlyForecastCard (month-to-date hero,
"≈$2.0k by month end" chip, the calibrated range bar, avg/active-day +
pace-vs-linear caption). No number appears twice on the page anymore,
and the page is one card shorter.

Includes the RangeBar Weather-geometry restyle (merged from
ui/rangebar-weather) so the strip and live-activity bars render in the
proposed style for the sign-off screenshot.
